feat: add telegram-style chat ui
