+++ client/src/pages/Chat.js
@@ -3,8 +3,7 @@ import { useAuth } from '../hooks/useAuth';
 import { useChat } from '../hooks/useChat';
 
 // Components
-import Sidebar from '../components/chat/Sidebar';
-import ChatWindow from '../components/chat/ChatWindow';
+import ChatLayout from '../components/chat/ChatLayout';
 import ProfileModal from '../components/modals/ProfileModal';
 import CreateGroupModal from '../components/modals/CreateGroupModal';
 import UserProfileModal from '../components/modals/UserProfileModal';
@@ -14,16 +13,11 @@ const Chat = () => {
   const { currentUser } = useAuth();
   const { selectedChat } = useChat();
   
-  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
   const [isProfileModalOpen, setIsProfileModalOpen] = useState(false);
   const [isCreateGroupModalOpen, setIsCreateGroupModalOpen] = useState(false);
   const [isUserProfileModalOpen, setIsUserProfileModalOpen] = useState(false);
   const [isGroupInfoModalOpen, setIsGroupInfoModalOpen] = useState(false);
   const [selectedUser, setSelectedUser] = useState(null);
-
-  const toggleMobileMenu = () => {
-    setIsMobileMenuOpen(!isMobileMenuOpen);
-  };
 
   const openProfileModal = () => {
     setIsProfileModalOpen(true);
@@ -61,35 +55,14 @@ const Chat = () => {
 
   return (
     <div className="h-full flex relative overscroll-none min-h-0">
-      {isMobileMenuOpen && (
-        <div
-          className="fixed inset-0 bg-black/40 z-20 md:hidden transition-opacity"
-          onClick={toggleMobileMenu}
-        ></div>
-      )}
-
-      {/* Sidebar */}
-      <Sidebar
-        isMobileMenuOpen={isMobileMenuOpen}
-        toggleMobileMenu={toggleMobileMenu}
-        openProfileModal={openProfileModal}
-        openCreateGroupModal={openCreateGroupModal}
-        openUserProfileModal={openUserProfileModal}
-      />
-
-      {/* Chat Window */}
-      <ChatWindow
-        toggleMobileMenu={toggleMobileMenu}
-        openUserProfileModal={openUserProfileModal}
-        openGroupInfoModal={openGroupInfoModal}
-      />
+      <ChatLayout />
 
       {/* Modals */}
       {isProfileModalOpen && (
-        <ProfileModal 
-          isOpen={isProfileModalOpen} 
-          onClose={closeProfileModal} 
-          user={currentUser} 
+        <ProfileModal
+          isOpen={isProfileModalOpen}
+          onClose={closeProfileModal}
+          user={currentUser}
         />
       )}
 
@@ -109,10 +82,10 @@ const Chat = () => {
       )}
 
       {isGroupInfoModalOpen && selectedChat && selectedChat.isGroupChat && (
-        <GroupInfoModal 
-          isOpen={isGroupInfoModalOpen} 
-          onClose={closeGroupInfoModal} 
-          chat={selectedChat} 
+        <GroupInfoModal
+          isOpen={isGroupInfoModalOpen}
+          onClose={closeGroupInfoModal}
+          chat={selectedChat}
         />
       )}
     </div>
